separating out upsert to create/update. removing try/excepts and updating permissions code to use permission manager

+++ src/citation/related_models/citation_project.py
@@ -1,8 +1,8 @@
-from django.db import models
 from django.contrib.contenttypes.fields import GenericRelation
+from django.db import models
+
 from researchhub_access_group.constants import ADMIN, EDITOR
 from researchhub_access_group.related_models.permission_model import Permission
-
 from user.models import Organization
 from user.related_models.user_model import User
 from utils.models import DefaultAuthenticatedModel
@@ -41,28 +41,16 @@ class CitationProject(DefaultAuthenticatedModel):
     """--- METHODS ---"""
 
     def add_editors(self, editor_ids=[]):
-        try:
-            for editor_id in editor_ids:
-                editor_exists = self.permissions.filter(
-                    access_type=EDITOR, user=editor_id
-                ).exists()
-                if not editor_exists:
-                    self.permissions.create(
-                        access_type=EDITOR, user=User.objects.get(id=editor_id)
-                    )
-            return True
-        except Exception as error:
-            return False
+        for editor_id in editor_ids:
+            editor_exists = self.permissions.has_editor_user(editor_id)
+            if not editor_exists:
+                self.permissions.create(access_type=EDITOR, user=editor_id)
+        return True
 
     def remove_editors(self, editor_ids):
-        try:
-            for editor_id in editor_ids:
-                self.permissions.filter(
-                    access_type=EDITOR, user=User.objects.get(id=editor_id)
-                ).all().delete()
-            return True
-        except Exception as error:
-            return False
+        for editor_id in editor_ids:
+            self.permissions.filter(access_type=EDITOR, user=editor_id).all().delete()
+        return True
 
     def get_current_user_has_access(self, user):
         org_has_user = self.organization.org_has_user(user)
@@ -72,11 +60,8 @@ class CitationProject(DefaultAuthenticatedModel):
         if self.is_public:
             return True
         else:
-            return self.permissions.filter(user=user).exists()
+            return self.permissions.has_user(user)
 
     def set_creator_as_admin(self):
-        try:
-            self.permissions.create(access_type=ADMIN, user=self.created_by)
-            return True
-        except Exception as error:
-            return False
+        self.permissions.create(access_type=ADMIN, user=self.created_by)
+        return True
